@@ -1,22 +1,21 @@
 using System;
 
-namespace ResultOfTask
+namespace ErrorHandling.QuerySyntax;
+
+public static class ParseResultExtensions
 {
-    public static class ParseResultExtensions
+    public static Result<int> ParseIntResult(this string s, string error = null)
     {
-        public static Result<int> ParseIntResult(this string s, string error = null)
-        {
-            int v;
-            return int.TryParse(s, out v)
-                ? v.AsResult()
-                : Result.Fail<int>(error ?? "Не число " + s);
-        }
-        public static Result<Guid> ParseGuidResult(this string s, string error = null)
-        {
-            Guid v;
-            return Guid.TryParse(s, out v)
-                ? v.AsResult()
-                : Result.Fail<Guid>(error ?? "Не GUID " + s);
-        }
+        int v;
+        return int.TryParse(s, out v)
+            ? v.AsResult()
+            : Result.Fail<int>(error ?? "Не число " + s);
+    }
+    public static Result<Guid> ParseGuidResult(this string s, string error = null)
+    {
+        Guid v;
+        return Guid.TryParse(s, out v)
+            ? v.AsResult()
+            : Result.Fail<Guid>(error ?? "Не GUID " + s);
     }
 }
