@@ -1,78 +1,81 @@
 ﻿using System;
 
-namespace ResultOfTask
+namespace ErrorHandling;
+
+public class None
 {
-    public class None
+    private None()
     {
-        private None()
-        {
-        }
     }
-    public struct Result<T>
+}
+
+public struct Result<T>
+{
+    public Result(string error, T value = default(T))
     {
-        public Result(string error, T value = default(T))
-        {
-            Error = error;
-            Value = value;
-        }
-        public string Error { get; }
-        internal T Value { get; }
-        public T GetValueOrThrow()
-        {
-            if (IsSuccess) return Value;
-            throw new InvalidOperationException($"No value. Only Error {Error}");
-        }
-        public bool IsSuccess => Error == null;
+        Error = error;
+        Value = value;
     }
 
-    public static class Result
+    public string Error { get; }
+    internal T Value { get; }
+
+    public T GetValueOrThrow()
     {
-        public static Result<T> AsResult<T>(this T value)
-        {
-            return Ok(value);
-        }
+        if (IsSuccess) return Value;
+        throw new InvalidOperationException($"No value. Only Error {Error}");
+    }
 
-        public static Result<T> Ok<T>(T value)
-        {
-            return new Result<T>(null, value);
-        }
+    public bool IsSuccess => Error == null;
+}
 
-        public static Result<T> Fail<T>(string e)
-        {
-            return new Result<T>(e);
-        }
+public static class Result
+{
+    public static Result<T> AsResult<T>(this T value)
+    {
+        return Ok(value);
+    }
 
-        public static Result<T> Of<T>(Func<T> f, string error = null)
-        {
-            try
-            {
-                return Ok(f());
-            }
-            catch (Exception e)
-            {
-                return Fail<T>(error ?? e.Message);
-            }
-        }
+    public static Result<T> Ok<T>(T value)
+    {
+        return new Result<T>(null, value);
+    }
 
-        public static Result<TOutput> Then<TInput, TOutput>(
-            this Result<TInput> input,
-            Func<TInput, TOutput> continuation)
-        {
-            throw new NotImplementedException();
-        }
+    public static Result<T> Fail<T>(string e)
+    {
+        return new Result<T>(e);
+    }
 
-        public static Result<TOutput> Then<TInput, TOutput>(
-            this Result<TInput> input,
-            Func<TInput, Result<TOutput>> continuation)
+    public static Result<T> Of<T>(Func<T> f, string error = null)
+    {
+        try
         {
-            throw new NotImplementedException();
+            return Ok(f());
         }
+        catch (Exception e)
+        {
+            return Fail<T>(error ?? e.Message);
+        }
+    }
 
-        public static Result<TInput> OnFail<TInput>(
-            this Result<TInput> input,
-            Action<string> handleError)
-        {
-            throw new NotImplementedException();
-        }
+    public static Result<TOutput> Then<TInput, TOutput>(
+        this Result<TInput> input,
+        Func<TInput, TOutput> continuation)
+    {
+        throw new NotImplementedException();
+    }
+
+    public static Result<TOutput> Then<TInput, TOutput>(
+        this Result<TInput> input,
+        Func<TInput, Result<TOutput>> continuation)
+    {
+        throw new NotImplementedException();
+    }
+
+    public static Result<TInput> OnFail<TInput>(
+        this Result<TInput> input,
+        Action<string> handleError)
+    {
+        throw new NotImplementedException();
     }
 }
